move decoration types to separate file
