+++ ext-src/helpers/GenerateDecorations.ts
@@ -1,23 +1,7 @@
 import * as vscode from 'vscode';
 import { Problem } from '../types';
 import { FixSuggestionCommandHandler } from '../commands';
-
-const removedTextBackgroundColor = new vscode.ThemeColor('diffEditor.removedTextBackground');
-const insertedTextBackgroundColor = new vscode.ThemeColor('diffEditor.insertedTextBackground');
-
-export const problemDecoration = vscode.window.createTextEditorDecorationType({
-  backgroundColor: removedTextBackgroundColor,
-  isWholeLine: true,
-  overviewRulerLane: 7,
-  overviewRulerColor: removedTextBackgroundColor,
-});
-
-export const suggestionDecoration = vscode.window.createTextEditorDecorationType({
-  backgroundColor: insertedTextBackgroundColor,
-  isWholeLine: true,
-  overviewRulerLane: 7,
-  overviewRulerColor: insertedTextBackgroundColor,
-});
+import { problemDecoration } from '../decorations';
 
 export function GenerateDecorations(
   results: Problem[],
